rename usersDevices => user-devices
