+++ admin/src/templates/export.html
@@ -15,7 +15,7 @@
     <a ui-sref="export.dhis" translate>DHIS2</a>
   </li>
   <li role="presentation" ui-sref-active="active">
-    <a ui-sref="export.usersDevices" translate>usersDevices</a>
+    <a ui-sref="export.user-devices" translate>User Devices</a>
   </li>
 </ul>
 <div class="tab-content" ui-view="tab" mm-auth="can_export_all"></div>
